+++ src/api/axios.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 
 const instance = axios.create({
   baseURL: "https://wikied-api.vercel.app/9-3/",
-  timeout: 1000,
+  // timeout: 1000,
 });
 
 export default instance;
