+++ src/components/PaginationBar.tsx
@@ -1,18 +1,34 @@
 interface PaginationBarProps {
   totalPage: number;
-  pageArr: number[];
   currentPage: number;
   handlePageChange: (page: number) => void;
   isLoading: boolean;
+  maxPage: number;
 }
 
 const PaginationBar = ({
   totalPage,
-  pageArr,
   currentPage,
   handlePageChange,
-  isLoading
+  isLoading,
+  maxPage = 5,
 }: PaginationBarProps) => {
+  let startPage;
+  let calNum;
+
+  if (totalPage <= maxPage) startPage = 1;
+  else {
+    calNum = Math.ceil(currentPage / maxPage);
+    startPage = (calNum - 1) * maxPage + 1;
+  }
+
+  const pageArr = Array.from(
+    {
+      length: Math.min(maxPage, totalPage),
+    },
+    (_, i) => startPage + i
+  );
+
   return (
     <div className="flex gap-[10px]">
       <button
